assignment4/requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
google-api-core==2.8.2
google-auth==2.9.0
google-cloud-core==2.3.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
idna==3.3
joblib==1.1.0
nltk==3.7
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
regex==2022.7.9
requests==2.28.1
rsa==4.8
six==1.16.0
tqdm==4.64.0
urllib3==1.26.10
